+++ app/persistence/app-dta-user.js
@@ -473,7 +473,7 @@ async function delUser( uid ) {
 async function getApiAppScopes( appId, appSecret ) {
   let authTbl = await getAuthTbl()
   let hash = createHash('sha256').update( appSecret ).digest('hex')
-  log.info( 'getApiAppScopes', hash,  authTbl[ appId ] )
+  // log.info( 'getApiAppScopes', hash,  authTbl[ appId ] )
   if ( authTbl[ appId ] && authTbl[ appId ].password == hash ) {
     return authTbl[ appId ].role.api
   } else {